major changes to input handling in preparation for pasting, which will require support for large blocks of input:
output which cannot be written to the tty (pty) is no longer
discarded. instead, the display module is responsible for keeping the
buffer alive until it has been entirely consumed. one simple (and
smart) way to ensure this is to refuse to process events that could
generate new input text as long as there is unwritten data. this works
as long as there is an earlier layer of buffering (for fbcon, kernel;
for x11, xlib). a direct-on-x-protocol display implementation will
have to do its own buffer management.

this architecture allows pasting to be performed direct from a buffer
owned by the display module, which could reside in static or shared
memory, or even as a memory-mapped file. it also facilitates reuse of
the code in non-tty-connected settings.