major internal changes in representation of character cells.
we now use 12 bytes per cell instead of 10. however, this allows us to
support 256-color mode (not yet implemented but the framework is in
place) and to mix scripts when using combining characters. while the
latter sounds ridiculous at first, being able to visibly see a
combining letter attached to a [, ", or ' is extremely useful in
scripting and regular expressions with some languages.

some code is left slightly messy, but overall it's much cleaner now
since struct uucell is now properly encapsulated.