additions to the decomposition table. these are _not_ u**code nfd equivalences but rather workarounds for the fact that the fools who added the greek extension block to unicode didn't know greek and didn't know they were duplicating existing characters. u**code stability policy makes it impossible to fix this stupid stupid mistake so i'll just maintain a table of corrections.
without these additional mappings, the duplicate characters will not
decompose and thus will show up as ?'s unless the font includes
precomposed glyphs for each of these broken compatibility characters.

in the future, further additions outside the standard may be made to
this table if the character charts indicate equivalences that cannot
be made canonical for stability reasons.